+++ samples/framework/internal/immediate.cc
@@ -31,9 +31,8 @@
 
 #include <cassert>
 
-#include "ozz/base/memory/allocator.h"
-
 #include "camera.h"
+#include "ozz/base/memory/allocator.h"
 #include "renderer_impl.h"
 #include "shader.h"
 
@@ -42,7 +41,7 @@ namespace sample {
 namespace internal {
 
 GlImmediateRenderer::GlImmediateRenderer(RendererImpl* _renderer)
-    : vbo_(0), buffer_(nullptr), max_size_(0), size_(0), renderer_(_renderer) {}
+    : vbo_(0), size_(0), renderer_(_renderer) {}
 
 GlImmediateRenderer::~GlImmediateRenderer() {
   assert(size_ == 0 && "Immediate rendering still in use.");
@@ -51,14 +50,10 @@ GlImmediateRenderer::~GlImmediateRenderer() {
     GL(DeleteBuffers(1, &vbo_));
     vbo_ = 0;
   }
-  ozz::memory::default_allocator()->Deallocate(buffer_);
-  buffer_ = nullptr;
 }
 
 bool GlImmediateRenderer::Initialize() {
   GL(GenBuffers(1, &vbo_));
-  const size_t kDefaultVboSize = 2 << 10;
-  ResizeVbo(kDefaultVboSize);
 
   immediate_pc_shader = ImmediatePCShader::Build();
   if (!immediate_pc_shader) {
@@ -80,7 +75,7 @@ template <>
 void GlImmediateRenderer::End<VertexPC>(GLenum _mode,
                                         const ozz::math::Float4x4& _transform) {
   GL(BindBuffer(GL_ARRAY_BUFFER, vbo_));
-  GL(BufferData(GL_ARRAY_BUFFER, size_, buffer_, GL_STREAM_DRAW));
+  GL(BufferData(GL_ARRAY_BUFFER, size_, buffer_.data(), GL_STREAM_DRAW));
 
   immediate_pc_shader->Bind(_transform, renderer_->camera()->view_proj(),
                             sizeof(VertexPC), 0, sizeof(VertexPC), 12);
@@ -100,7 +95,7 @@ template <>
 void GlImmediateRenderer::End<VertexPTC>(
     GLenum _mode, const ozz::math::Float4x4& _transform) {
   GL(BindBuffer(GL_ARRAY_BUFFER, vbo_));
-  GL(BufferData(GL_ARRAY_BUFFER, size_, buffer_, GL_STREAM_DRAW));
+  GL(BufferData(GL_ARRAY_BUFFER, size_, buffer_.data(), GL_STREAM_DRAW));
 
   immediate_ptc_shader->Bind(_transform, renderer_->camera()->view_proj(),
                              sizeof(VertexPTC), 0, sizeof(VertexPTC), 12,
@@ -115,14 +110,6 @@ void GlImmediateRenderer::End<VertexPTC>(
 
   // Reset vertex count for the next call
   size_ = 0;
-}
-
-void GlImmediateRenderer::ResizeVbo(size_t _new_size) {
-  if (_new_size > max_size_) {
-    max_size_ = ozz::math::Max(max_size_ * 2, _new_size);
-    buffer_ = reinterpret_cast<char*>(
-        ozz::memory::default_allocator()->Reallocate(buffer_, max_size_, 16));
-  }
 }
 }  // namespace internal
 }  // namespace sample